AI agent of New Year Painting/年画作品个性化推荐大模型代码/250929-训练模型/requirements.txt
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.0
torch==2.0.1
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.0.0
openai==0.27.8
langchain==0.0.266
faiss-cpu==1.7.4
transformers==4.31.0
accelerate==0.21.0
diffusers==0.19.0
flask==2.3.2
flask-cors==4.0.0
uvicorn==0.23.1
fastapi==0.100.0
python-multipart==0.0.6
dotenv==1.0.0
deep_translator==1.11.4
pypinyin==0.49.0